requirements.txt
python-dotenv
rumps
wxPython
requests
Flask
Flask-CORS
anthropic
